Add a tools/test script.
The script can be used to test the output from Java on the given
MiniJava source file, or the output from compiling with mjc and
then running on a remote ARM host, or it can do both and then
compare the result.

Thought this script may come in handy in the future.

Also modified the ARMMain class to print the ARM assembly code
for printing "42" by default. It helped me test that the script
works like it should.

Script info:

Usage: test java <infile> [class]
       test mjc <infile> <host>
       test compare <infile> <host> [class]

Available subcommands:
   java     Compile/run <infile> locally using Java.
   mjc      Compile <infile> using mjc, then assemble/run on <host>.
   compare  Run above commands and compare results.

The optional class parameter specifies which class to run when
invoking the Java VM. If unspecified, it will be deduced from the
input file name