Add tools/submit script and a "submit" ANT target to invoke it.
The script takes mjc.tar.gz, encodes it using uuencode and sends
it to the specified e-mail address (default: submit@tigris.csc.kth.se)
using the mail command.

The "submit" ANT target simply invokes the script. It depends on the
"tar" target to generate the .tar.gz, which in turn requires that
either the

    -Dtigris.id=MY_EMAIL
    -Dtigris.email=MY_EMAIL

options are passed to ANT, or that the TIGRIS_EMAIL and TIGRIS_ID
environment variables are set. The ANT options override the environ-
ment variables.

By default, the "submit" target passes submit@tigris.csc.kth.se as
the recipient address to the script. This can be changed using the
-Dtigris.to=<TO_EMAIL> option.